requirements.txt
# ─── Core Agentic Framework ───────────────────────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langgraph>=0.1.0

# ─── LLM Providers ────────────────────────────────────────────────────────────
openai>=1.30.0
anthropic>=0.28.0

# ─── RAG / Memory ─────────────────────────────────────────────────────────────
chromadb>=0.5.0
langchain-chroma>=0.1.0
tiktoken>=0.7.0

# ─── Web / HTTP Tools ─────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.32.0
beautifulsoup4>=4.12.0

# ─── ZAP Integration ──────────────────────────────────────────────────────────
python-owasp-zap-v2.4>=0.0.21

# ─── Nuclei (subprocess-based, install nuclei binary separately) ──────────────
# nuclei binary: https://github.com/projectdiscovery/nuclei/releases

# ─── Data / Parsing ───────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
rich>=13.7.0
jinja2>=3.1.4

# ─── Async ────────────────────────────────────────────────────────────────────
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# ─── Reporting ────────────────────────────────────────────────────────────────
markdown>=3.6
weasyprint>=62.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
